refactor!: Rename `memory_output_stream("2"->"").hpp`
The old filename is a left over from the coexistence with the legacy
streams.

Refactor all test code to use the test output streams instead the public
`memory_output_stream` API.

BREAKING CHANGE: Includes of `dplx/dp/streams/memory_output_stream2.hpp`
        must be updated to `dplx/dp/streams/memory_output_stream.hpp`
