+++ src/dplx/dp/codecs/core.test.cpp
@@ -13,7 +13,6 @@
 #include <dplx/cncr/misc.hpp>
 
 #include "dplx/dp/api.hpp"
-#include "dplx/dp/streams/memory_output_stream2.hpp"
 #include "item_sample_ct.hpp"
 #include "test_input_stream.hpp"
 #include "test_output_stream.hpp"
@@ -75,8 +74,7 @@ TEST_CASE("bool is encodable")
 
     SECTION("with encode")
     {
-        std::vector<std::byte> encodingBuffer(sample.encoded_length);
-        dp::memory_output_stream outputStream(encodingBuffer);
+        simple_test_output_stream outputStream(sample.encoded_length);
 
         REQUIRE(dp::encode(outputStream, sample.value));
 
@@ -108,8 +106,7 @@ TEST_CASE("float has a codec")
 
     SECTION("with encode")
     {
-        std::vector<std::byte> encodingBuffer(sample.encoded_length);
-        dp::memory_output_stream outputStream(encodingBuffer);
+        simple_test_output_stream outputStream(sample.encoded_length);
 
         REQUIRE(dp::encode(outputStream, sample.value));
 
@@ -141,8 +138,7 @@ TEST_CASE("double has a codec")
 
     SECTION("with encode")
     {
-        std::vector<std::byte> encodingBuffer(sample.encoded_length);
-        dp::memory_output_stream outputStream(encodingBuffer);
+        simple_test_output_stream outputStream(sample.encoded_length);
 
         REQUIRE(dp::encode(outputStream, sample.value));
