+++ src/dplx/dp/streams/memory_output_stream.hpp
